Move JsConfig and WasmConfig to build-logic
