+++ build-logic/src/main/kotlin/ktorbuild/NamedDomainObjectCollection.kt
@@ -1,0 +1,11 @@
+/*
+ * Copyright 2014-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package ktorbuild
+
+import org.gradle.api.NamedDomainObjectCollection
+
+internal fun <T> NamedDomainObjectCollection<T>.maybeNamed(name: String, configure: T.() -> Unit) {
+    if (name in names) named(name).configure(configure)
+}
