+++ build-logic/src/main/kotlin/ktorbuild/internal/String.kt
@@ -1,0 +1,7 @@
+/*
+ * Copyright 2014-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package ktorbuild.internal
+
+internal fun String.capitalized() = replaceFirstChar { it.uppercase() }
