--- buildSrc/src/main/kotlin/JsConfig.kt
@@ -1,34 +1,0 @@
-/*
- * Copyright 2014-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
- */
-
-import internal.capitalized
-import internal.libs
-import org.gradle.api.Project
-import org.gradle.kotlin.dsl.invoke
-
-fun Project.configureJs() {
-    kotlin {
-        js {
-            binaries.library()
-        }
-
-        sourceSets {
-            jsTest {
-                dependencies {
-                    implementation(npm("puppeteer", libs.versions.puppeteer.get()))
-                }
-            }
-        }
-    }
-
-    configureJsTestTasks(target = "js")
-}
-
-internal fun Project.configureJsTestTasks(target: String) {
-    val shouldRunJsBrowserTest = !hasProperty("teamcity") || hasProperty("enable-js-tests")
-    if (shouldRunJsBrowserTest) return
-
-    tasks.maybeNamed("clean${target.capitalized()}BrowserTest") { onlyIf { false } }
-    tasks.maybeNamed("${target}BrowserTest") { onlyIf { false } }
-}
